".pyc" files are now no longer tracked by git.
These should be created by python from the source code automatically when
the code is run. I suspect that having them in the repository is superfluous,
but I don't claim to be an expert on what .pyc files do or whether it is
appropriate to store them in a git repository.